+++ src/components/exercise-scenarios/ExerciseScenarios.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import AppPage from "../shared/AppPage";
-import scenarios from "./../../dummy-data/exercise-scenarios"
+import AppPage from '../shared/AppPage';
+import {exerciseScenarioById, scenarios} from '../../dummy-data/exercise-scenarios'
+import { getSorting, stableSort } from '../../helpers/table-sort-helper'
 import Paper from '@material-ui/core/Paper';
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
@@ -10,32 +11,6 @@ import TableRow from '@material-ui/core/TableRow';
 import Tooltip from "@material-ui/core/Tooltip/Tooltip";
 import TableSortLabel from "@material-ui/core/TableSortLabel/TableSortLabel";
 
-function desc(a, b, orderBy) {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
-  }
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-  return 0;
-}
-
-function stableSort(array, cmp) {
-  const stabilizedThis = array.map((el, index) => [el, index]);
-  stabilizedThis.sort((a, b) => {
-    const order = cmp(a[0], b[0]);
-    if (order !== 0) return order;
-    return a[1] - b[1];
-  });
-  return stabilizedThis.map(el => el[0]);
-}
-
-function getSorting(order, orderBy) {
-  return order === "desc"
-    ? (a, b) => desc(a, b, orderBy)
-    : (a, b) => -desc(a, b, orderBy);
-}
-
 const rows = [
   { id: "week", numeric: true, label: "Uke" },
   { id: "type", numeric: false, label: "Type" },
@@ -43,7 +18,7 @@ const rows = [
   { id: "scenario", numeric: false, label: "Scenario" },
 ];
 
-class EnhancedTableHead extends React.Component {
+class ExerciseScenariosTableHead extends React.Component {
   createSortHandler = property => event => {
     this.props.onRequestSort(event, property);
   };
@@ -59,7 +34,6 @@ class EnhancedTableHead extends React.Component {
               <TableCell
                 key={row.id}
                 numeric={row.numeric}
-                padding={row.disablePadding ? "none" : "default"}
                 sortDirection={orderBy === row.id ? order : false}
               >
                 <Tooltip
@@ -84,12 +58,16 @@ class EnhancedTableHead extends React.Component {
   }
 }
 
-export default class EnhancedTable extends React.Component {
-  state = {
-    order: "asc",
-    orderBy: "week",
-    data: scenarios,
-  };
+export default class ExerciseScenarios extends React.Component {
+
+  constructor(props) {
+    super(props);
+    this.state = {
+      order: "asc",
+      orderBy: "week",
+      data: scenarios,
+    }
+  }
 
   handleRequestSort = (event, property) => {
     const orderBy = property;
@@ -101,37 +79,19 @@ export default class EnhancedTable extends React.Component {
   };
 
   handleClick = (event, id) => {
-    const { selected } = this.state;
-    const selectedIndex = selected.indexOf(id);
-    let newSelected = [];
-
-    if (selectedIndex === -1) {
-      newSelected = newSelected.concat(selected, id);
-    } else if (selectedIndex === 0) {
-      newSelected = newSelected.concat(selected.slice(1));
-    } else if (selectedIndex === selected.length - 1) {
-      newSelected = newSelected.concat(selected.slice(0, -1));
-    } else if (selectedIndex > 0) {
-      newSelected = newSelected.concat(
-        selected.slice(0, selectedIndex),
-        selected.slice(selectedIndex + 1)
-      );
-    }
-
-    this.setState({ selected: newSelected });
+    const selected = exerciseScenarioById[id];
+    console.log("Scenario " + selected.id + " is selected.")
   };
 
   render() {
-    const { data, order, orderBy, rowsPerPage, page } = this.state;
-    const emptyRows =
-      rowsPerPage - Math.min(rowsPerPage, data.length - page * rowsPerPage);
+    const { data, order, orderBy } = this.state;
 
     return (
       <AppPage title="Øvelsesscenarioer" back="..">
         <Paper>
           <div>
             <Table>
-              <EnhancedTableHead
+              <ExerciseScenariosTableHead
                 order={order}
                 orderBy={orderBy}
                 onRequestSort={this.handleRequestSort}
@@ -155,11 +115,6 @@ export default class EnhancedTable extends React.Component {
                       </TableRow>
                     );
                   })}
-                {emptyRows > 0 && (
-                  <TableRow style={{ height: 49 * emptyRows }}>
-                    <TableCell colSpan={6} />
-                  </TableRow>
-                )}
               </TableBody>
             </Table>
           </div>
